Refactor import paths for resolvers and enums
