+++ src/app/app.module.ts
@@ -11,7 +11,7 @@ import {
 } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { httpErrorInterceptor } from '@interceptors';
-import { AppInitializerService } from '@services';
+import { AppInitializerService } from '@services/app-initializer-service/app-initializer.service';
 import { MojBannerComponent, MojHeaderComponent, MojHeaderNavigationItemComponent } from '@components/moj';
 import { GovukFooterComponent } from '@components/govuk';
