+++ src/app/flows/fines/fines-mac/fines-mac-create-account/fines-mac-create-account-form/fines-mac-create-account-form.component.ts
@@ -22,7 +22,7 @@ import {
 } from '@components/govuk';
 
 import { AbstractFormBaseComponent } from '@components/abstract';
-import { RoutingPaths } from '@enums/routing-paths';
+import { RoutingPaths } from '@routing/enums/routing-paths';
 import { IFinesMacCreateAccountAccountTypes } from '../interfaces/fines-mac-create-account-account-types.interface';
 import { IFinesMacCreateAccountControlNames } from '../interfaces/fines-mac-create-account-control-names.interface';
 import { IFinesMacCreateAccountFieldErrors } from '../interfaces/fines-mac-create-account-field-errors.interface';
