+++ src/app/guards/route-permissions/route-permissions.guard.spec.ts
@@ -20,7 +20,7 @@ import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http'
 
 import { FINES_ROUTING_PATHS } from '@routing/fines/constants/fines-routing-paths.constant';
 import { FINES_ROUTING_PERMISSIONS } from '@routing/fines/constants/fines-routing-permissions.constant';
-import { RoutingPaths } from '@enums/routing-paths';
+import { RoutingPaths } from '@routing/enums/routing-paths';
 import { IFinesRoutingPermissions } from '@routing/fines/interfaces/fines-routing-permissions.interface';
 
 async function runRoutePermissionGuard(
