+++ src/app/pages/sign-in/sign-in.component.ts
@@ -5,7 +5,7 @@ import { SignInSsoComponent } from './sign-in-sso/sign-in-sso.component';
 import { SignInStubComponent } from './sign-in-stub/sign-in-stub.component';
 import { GlobalStateService } from '@services/global-state-service/global-state.service';
 import { ISignInStubForm } from './interfaces';
-import { SsoEndpoints } from '@enums/sso-endpoints';
+import { SsoEndpoints } from '@routing/enums/sso-endpoints';
 
 @Component({
   selector: 'app-sign-in',
