--- src/app/resolvers/index.ts
@@ -1,3 +1,0 @@
-import { userStateResolver } from './user-state/user-state.resolver';
-
-export { userStateResolver };
