+++ src/app/routing/enums/sso-endpoints.ts
