+++ src/app/services/auth-service/auth.service.ts
@@ -2,7 +2,7 @@ import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable, inject } from '@angular/core';
 import { catchError, tap, throwError } from 'rxjs';
 
-import { SsoEndpoints } from '../../enums/sso-endpoints';
+import { SsoEndpoints } from '@routing/enums/sso-endpoints';
 import { GlobalStateService } from '../global-state-service/global-state.service';
 
 @Injectable({
