+++ src/app/services/session-service/enums/session-endpoints.ts
