+++ src/app/services/session-service/session.service.ts
@@ -1,6 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable, inject } from '@angular/core';
-import { SessionEndpoints } from '@enums/session-endpoints';
+import { SessionEndpoints } from '@services/session-service/enums/session-endpoints';
 import { ISessionTokenExpiry } from '@services/session-service/interfaces/session-token-expiry.interface';
 import { ISessionUserState } from '@services/session-service/interfaces/session-user-state.interface';
 import { GlobalStateService } from '@services/global-state-service/global-state.service';
